Refactor form-related code into a separate class
