+++ includes/class-sample-spectrum-api.php
@@ -28,11 +28,12 @@ class Sample_Spectrum_API {
 	public $client_id;
 
 	/**
-	 * Setup $client_id.
+	 * Setup $client_id ($api_key is ignored and present only for compatibility).
 	 *
 	 * @param string $client_id Client ID.
+	 * @param string $api_key   API Key.
 	 */
-	public function __construct( $client_id ) {
+	public function __construct( $client_id, $api_key=null ) {
 		$this->client_id = $client_id;
 	}
